rename streaming.tsx to ParentComponent.tsx
